% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OkeeffeColor.R
\docType{data}
\name{okeeffe.col.info}
\alias{okeeffe.col.info}
\title{Color plaettes information}
\format{An object of class \code{data.frame} with 3 rows and 3 columns.}
\usage{
okeeffe.col.info
}
\description{
Use \code{okeeffe.col.info} to check information of available color plaettes.
}
\keyword{datasets}
