% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OkeeffeColor.R
\name{okeeffe.col.show}
\alias{okeeffe.col.show}
\title{Show a specific color palette}
\usage{
okeeffe.col.show(name, n = NULL)
}
\arguments{
\item{name}{Name of a specific plaette. Name can be one of the following: \code{Black_Iris_II}, \code{Blue_and_Green_Music}, \code{Two_Pink_Shells}.}

\item{n}{Color number of a specific plaette. n should be larger than 1 and smaller than the maximum color number of the specific palette. If omitted, the maximum color number of the specific palette will be used.}
}
\value{
An image showing a specific plaette.
}
\description{
Visually show a specific color palette from Georgia O'Keeffe's painting.
}
\examples{
okeeffe.col.show("Blue_and_Green_Music")
okeeffe.col.show("Black_Iris_II", 3)
okeeffe.col.show("Two_Pink_Shells", 2)
}
